Implement client to use multiple Brooker. Change Channel to reall class
Before sending message to a topic, in 0.8 the client need to determine
the kafka broker leader. The Channel is no longer an abstract class
to accommodate efficient connection between multiple broker
(so that it will not create multiple connection to same broker).

As a side effect, the Channel also change visibility of function.